can't delete an internship of a different company anymore

+++ src/resolver/DeleteInternshipOffer.ts
@@ -7,14 +7,18 @@ export const deleteInternship = async ({input}, ctx: Context) => {
     }
 
     if (ctx.user.role !== 'company') {
-        throw new UserError("Can't register an internship for a non-existing company");
+        throw new UserError("Can't delete an internship for a non-existing company");
     }
-    console.log(input);
+
     const internshipRepository = ctx.repositoryFactory.getInternshipRepository();
     const internship = await internshipRepository.getById(input);
-    console.log(internship);
     if (internship === null) {
         throw new UserError("Can't delete a not existing internship");
+    }
+    const companyRepository = ctx.repositoryFactory.getCompanyRepository();
+    const company = await companyRepository.getCompanyByUsername(ctx.user.username);
+    if (company.id !== internship.company) {
+        throw new UserError("Can't delete an internship of a different company.");
     }
 
     return await internshipRepository.remove(internship);
